build(ingest): refresh the pixi lock
Re-solves the conda environment to current: aws-c-* , libabseil, libprotobuf
6 -> 7, and the macOS floor moves to 12.0.

The ruff that comes with it (0.16) adds RUF046, which correctly flags the
int(round(...)) in tides.py — round() on a float already returns an int.

Verified: pixi install, ruff format + check clean, every ingest module imports
and the buoy/tide/wind registries resolve on the new environment.

+++ ingest/tides.py
@@ -266,7 +266,7 @@ def refresh_port(
     win_start, win_end = int(start.timestamp()), int(end.timestamp())
     new_df = pl.DataFrame(
         {
-            "t": [int(round(e["t"])) for e in new_events],
+            "t": [round(e["t"]) for e in new_events],
             "h": [e["h"] for e in new_events],
             "k": [e["k"] for e in new_events],
         },
